+++ Core/Models/Enums/SideEnum.cs
@@ -1,6 +1,7 @@
 ﻿namespace BN.PROJECT.Core;
 public enum SideEnum
 {
+    None,
     Buy,
     Sell
 }