+++ StrategyService/Domain/Services/Contracts/IStrategyService.cs
@@ -3,7 +3,8 @@
 public interface IStrategyService
 {
     Task StartTest(StrategyMessage message);
-    Task EvaluateQuote(Guid testId, Quote quote);
+    Task EvaluateQuote(Guid strategyId, Quote quote);
     Task StopTest(StrategyMessage message);
+    List<Position>? GetPositions(Guid strategyId);
     bool CanHandle(StrategyEnum strategy);
 }