+++ StrategyService/Domain/Services/Implementation/Strategies/BreakoutStrategy.cs
@@ -1,4 +1,5 @@
-﻿using System.Diagnostics;
+﻿using NuGet.Configuration;
+using System.Diagnostics;
 
 namespace BN.PROJECT.StrategyService;
 
@@ -28,7 +29,7 @@ public class BreakoutStrategy : IStrategyService
             return Task.CompletedTask;
         }
         var strategySettings = message.Settings;
-        var breakOutSettings = _strategyOperations.GetBreakoutModel(strategySettings);
+        var breakOutSettings = JsonConvert.DeserializeObject<BreakoutModel>(strategySettings.StrategyParams);
 
         var tpm = new BreakoutProcessModel
         {
@@ -50,6 +51,7 @@ public class BreakoutStrategy : IStrategyService
             TakeProfitPercent = strategySettings.TakeProfitPercent,
             StopLossPercent = strategySettings.StopLossPercent,
             TrailingStop = strategySettings.TrailingStop,
+            StopLossType = breakOutSettings.StopLossType,
         };
         _ = _strategyProcesses.TryAdd(message.StrategyId, tpm);
         _ = _positions.TryAdd(message.StrategyId, new List<Position>());
@@ -57,11 +59,11 @@ public class BreakoutStrategy : IStrategyService
     }
 
 
-    public Task EvaluateQuote(Guid testId, Quote quote)
+    public Task EvaluateQuote(Guid strategyId, Quote quote)
     {
         var stopwatch = Stopwatch.StartNew();
 
-        if (!_strategyProcesses.ContainsKey(testId))
+        if (!_strategyProcesses.ContainsKey(strategyId))
         {
             return Task.CompletedTask;
         }
@@ -72,7 +74,7 @@ public class BreakoutStrategy : IStrategyService
         }
 
 
-        var tpm = _strategyProcesses[testId];
+        var tpm = _strategyProcesses[strategyId];
         var quoteStamp = quote.TimestampUtc;
         if (tpm.StartDate == DateTime.MinValue)
         {
@@ -132,7 +134,7 @@ public class BreakoutStrategy : IStrategyService
 
 
 
-        var openPosition = _positions[testId].Where(p => p.PriceClose == 0).FirstOrDefault();
+        var openPosition = _positions[strategyId].Where(p => p.PriceClose == 0).FirstOrDefault();
         if (openPosition == null)
         {
             var breakoutParams = JsonConvert.SerializeObject(new
@@ -146,38 +148,40 @@ public class BreakoutStrategy : IStrategyService
             // check breakout high
             if (quote.AskPrice > tpm.PrevHigh && quote.AskPrice - tpm.PrevHigh < 1)
             {
+                var sl = tpm.StopLossType == StopLossTypeEnum.None ? quote.BidPrice + (quote.BidPrice * tpm.StopLossPercent / 100) : tpm.PrevLow;
                 var position = PositionExtensions.CreatePosition(
                                      tpm.StrategyId,
                                      tpm.Asset,
                                      tpm.Quantity,
                                      SideEnum.Buy,
                                      quote.AskPrice,
-                                     tpm.PrevLow,
+                                     sl,
                                      quote.BidPrice + (quote.BidPrice * tpm.TakeProfitPercent / 100),
                                      quote.TimestampUtc,
                                      StrategyEnum.Breakout,
                                      breakoutParams);
 
 
-                _positions[testId].Add(position);
+                _positions[strategyId].Add(position);
             }
 
             // check breakout low
             if (quote.BidPrice < tpm.PrevLow && tpm.PrevLow - quote.BidPrice < 1)
             {
+                var sl = tpm.StopLossType == StopLossTypeEnum.None ? quote.AskPrice - (quote.AskPrice * tpm.StopLossPercent / 100) : tpm.PrevHigh;
                 var position = PositionExtensions.CreatePosition(
                                    tpm.StrategyId,
                                    tpm.Asset,
                                    tpm.Quantity,
                                    SideEnum.Sell,
                                    quote.AskPrice,
-                                   tpm.PrevHigh,
+                                   sl,
                                    quote.BidPrice - (quote.BidPrice * tpm.TakeProfitPercent / 100),
                                    quote.TimestampUtc,
                                    StrategyEnum.Breakout,
                                    breakoutParams);
 
-                _positions[testId].Add(position);
+                _positions[strategyId].Add(position);
             }
         }
         else
@@ -187,62 +191,23 @@ public class BreakoutStrategy : IStrategyService
             {
                 var closePrice = openPosition.Side == SideEnum.Buy ? quote.BidPrice : quote.AskPrice;
                 openPosition.ClosePosition(quote.TimestampUtc, closePrice, "EoD Close");
-                return Task.CompletedTask;
             }
-            if (openPosition.Side == SideEnum.Buy)
+            else
             {
-                if (quote.BidPrice > openPosition.TakeProfit)
-                {
-                    if (tpm.TrailingStop > 0m)
-                    {
-                        var tp = quote.BidPrice + (quote.BidPrice * tpm.TakeProfitPercent / 100);
-                        var sl = quote.BidPrice - (quote.BidPrice * tpm.TrailingStop / 100);
-                        openPosition.UpdateTakeProfitAndStopLoss(tp, sl);
-                    }
-                    else
-                    {
-                        openPosition.ClosePosition(quote.TimestampUtc, quote.BidPrice, "TP");
-                    }
-                }
-
-                if (quote.BidPrice < openPosition.StopLoss)
-                {
-                    openPosition.ClosePosition(quote.TimestampUtc, quote.BidPrice, "SL");
-                }
-            }
-
-            if (openPosition.Side == SideEnum.Sell)
-            {
-                if (quote.AskPrice < openPosition.TakeProfit)
-                {
-                    if (tpm.TrailingStop > 0m)
-                    {
-                        var tp = quote.AskPrice - (quote.AskPrice * tpm.TakeProfitPercent / 100);
-                        var sl = quote.AskPrice + (quote.AskPrice * tpm.TrailingStop / 100);
-                        openPosition.UpdateTakeProfitAndStopLoss(tp, sl);
-                    }
-                    else
-                    {
-                        openPosition.ClosePosition(quote.TimestampUtc, quote.AskPrice, "TP");
-                    }
-                }
-
-                if (quote.AskPrice > openPosition.StopLoss)
-                {
-                    openPosition.ClosePosition(quote.TimestampUtc, quote.AskPrice, "SL");
-                }
+                _strategyOperations.UpdateOrCloseOpenPosition(ref openPosition, quote, tpm.TrailingStop, tpm.TakeProfitPercent);
             }
         }
         stopwatch.Stop();
         _logger.LogInformation($"Breakout EvaluateQuote done in {stopwatch.Elapsed.TotalNanoseconds} ns");
         return Task.CompletedTask;
     }
+
     public async Task StopTest(StrategyMessage message)
     {
         _logger.LogInformation("Backtest stopped");
-        var testId = message.StrategyId;
+        var strategyId = message.StrategyId;
 
-        var pos = _positions[testId].ToList();
+        var pos = _positions[strategyId].ToList();
 
         var overNight = pos.Where(p => p.StampOpened.Day != p.StampClosed.Day).ToList();
         var profits = pos.Where(p => p.ProfitLoss > 0).ToList();
@@ -267,12 +232,15 @@ public class BreakoutStrategy : IStrategyService
             await strategyRepository.AddPositionsAsync(pos);
         }
 
+        _strategyProcesses.TryRemove(strategyId, out _);
+        _positions.TryRemove(strategyId, out _);
+        _logger.LogInformation($"#BT {strategyId} Test stopped");
 
-
-        _strategyProcesses.TryRemove(testId, out _);
-        _positions.TryRemove(testId, out _);
-        _logger.LogInformation($"#BT {testId} Test stopped");
-
+    }
+    public List<Position>? GetPositions(Guid strategyId)
+    {
+        var positions = _positions.ContainsKey(strategyId) ? _positions[strategyId] : null;
+        return positions;
     }
 
     public bool CanHandle(StrategyEnum strategy) =>