+++ StrategyService/Domain/Services/Implementation/StrategyOperations.cs
@@ -26,9 +26,50 @@ public class StrategyOperations : IStrategyOperations
             _ => TimeSpan.FromDays(1)
         };
     }
-
-    public BreakoutModel GetBreakoutModel(StrategySettingsModel settings)
+    public void UpdateOrCloseOpenPosition(ref Position openPosition, Quote quote, decimal trailingStop, decimal takeProfitPercent )
     {
-        return JsonConvert.DeserializeObject<BreakoutModel>(settings.StrategyParams);
+        if (openPosition.Side == SideEnum.Buy)
+        {
+            if (quote.BidPrice > openPosition.TakeProfit)
+            {
+                if (trailingStop > 0m)
+                {
+                    var tp = quote.BidPrice + (quote.BidPrice * takeProfitPercent / 100);
+                    var sl = quote.BidPrice - (quote.BidPrice * trailingStop / 100);
+                    openPosition.UpdateTakeProfitAndStopLoss(tp, sl);
+                }
+                else
+                {
+                    openPosition.ClosePosition(quote.TimestampUtc, quote.BidPrice, "TP");
+                }
+            }
+
+            if (quote.BidPrice < openPosition.StopLoss)
+            {
+                openPosition.ClosePosition(quote.TimestampUtc, quote.BidPrice, "SL");
+            }
+        }
+
+        if (openPosition.Side == SideEnum.Sell)
+        {
+            if (quote.BidPrice < openPosition.TakeProfit)
+            {
+                if (trailingStop > 0m)
+                {
+                    var tp = quote.BidPrice - (quote.BidPrice * takeProfitPercent / 100);
+                    var sl = quote.AskPrice + (quote.AskPrice * trailingStop / 100);
+                    openPosition.UpdateTakeProfitAndStopLoss(tp, sl);
+                }
+                else
+                {
+                    openPosition.ClosePosition(quote.TimestampUtc, quote.AskPrice, "TP");
+                }
+            }
+
+            if (quote.AskPrice > openPosition.StopLoss)
+            {
+                openPosition.ClosePosition(quote.TimestampUtc, quote.AskPrice, "SL");
+            }
+        }
     }
 }