+++ AlpacaService/Release-Notes.md
@@ -9,6 +9,11 @@ BN.PROJECT is a platform which combines brokerage operations with algorithmic tr
 - implement code coverage tool coverlet
 - implement code coverage report generator 
 
+## Version 2024.12.1
+
+- SMA Strategy Refactoring
+- Strategy Integration Tests
+
 ## Version 2024.11.2
 
 - sma strategy implementation